Extract hunt
Separate calculateDistance into another function called hunt which deals
with all ghosts hunting mechanics and call this function inside
calculateDistance. This is where the conditional will go which will
dictate whether the hunting or scattering movement will initiate.
